+++ app/controllers/query_controller.py
@@ -2,8 +2,10 @@ import motor.motor_asyncio as motor
 from litestar import Litestar, get
 from litestar.openapi.config import OpenAPIConfig
 
+from app.controllers.query.CommentController import CommentController
 from app.controllers.query.EventController import EventController
 from app.controllers.query.PostController import PostController
+from app.controllers.query.TagController import TagController
 from app.controllers.query.UserController import UserController, jwt_auth
 
 local_connection_string = ""
@@ -17,19 +19,19 @@ except Exception as e:
 
 mongo_db = client["social_db"]
 
-
-@get("/")
-async def hello_world() -> str:
-    return "Hello, Query!"
-
-
 openapi_config = OpenAPIConfig(
     title="Query API",
     version="1.0.0",
 )
 
 app = Litestar(
-    route_handlers=[hello_world, UserController, PostController, EventController],
+    route_handlers=[
+        UserController,
+        PostController,
+        EventController,
+        TagController,
+        CommentController,
+    ],
     on_app_init=[jwt_auth.on_app_init],
     openapi_config=openapi_config,
 )